Move structure manifest code into asset_management
Required to be loaded before StructureHandles, so must be defined there.
