+++ emergence_lib/src/asset_management/manifest/emergence_markers.rs
@@ -2,12 +2,11 @@
 
 use crate::{
     items::{recipe::RecipeData, ItemData},
-    structures::StructureData,
     terrain::TerrainData,
     units::UnitData,
 };
 
-use super::Manifest;
+use super::{structure::StructureData, Manifest};
 
 /// The marker type for [`Id<Recipe>`](super::Id).
 pub struct Recipe;
