+++ emergence_lib/src/asset_management/manifest/plugin.rs
@@ -9,7 +9,7 @@ use crate::asset_management::{AssetCollectionExt, AssetState, Loadable};
 use super::{
     loader::RawManifestLoader,
     raw::{RawItemManifest, RawManifest},
-    Manifest,
+    Manifest, StructureManifest,
 };
 
 /// A plugin to handle the creation of all manifest resources.
@@ -17,7 +17,8 @@ pub struct ManifestPlugin;
 
 impl Plugin for ManifestPlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugin(RawManifestPlugin::<RawItemManifest>::new())
+        app.init_resource::<StructureManifest>()
+            .add_plugin(RawManifestPlugin::<RawItemManifest>::new())
             // This is needed to ensure that the manifest resources are actually created in time for AssetState::Ready
             .add_system(
                 apply_system_buffers
